Change type of numSectors

+++ dmg/io.c
@@ -33,7 +33,7 @@ typedef struct {
 	// Read
 	pthread_mutex_t inMut;
 	AbstractFile* in;
-	uint32_t sectorsRemain;
+	off_t sectorsRemain;
 	uint32_t sectorsRead;
 	uint32_t curRun;
 	uint64_t curSector;
@@ -218,7 +218,7 @@ static void *threadWorker(void* arg) {
 	return NULL;
 }
 
-BLKXTable* insertBLKX(AbstractFile* out_, AbstractFile* in_, uint32_t firstSectorNumber, uint32_t numSectors_, uint32_t blocksDescriptor,
+BLKXTable* insertBLKX(AbstractFile* out_, AbstractFile* in_, uint32_t firstSectorNumber, off_t numSectors_, uint32_t blocksDescriptor,
 			uint32_t checksumType, ChecksumFunc uncompressedChk_, void* uncompressedChkToken_, ChecksumFunc compressedChk_,
 			void* compressedChkToken_, Volume* volume, AbstractAttribution* attribution_, Compressor* comp, size_t runSectors) {
 	threadData td = {
